dom/bindings/test/test_jsimplemented_subclassing.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1400275
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1400275</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1400275 **/
    class Foo extends RTCPeerConnection {
    }

    var obj = new Foo();
    is(Object.getPrototypeOf(obj), Foo.prototype,
       "Should have the right prototype");
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1400275">Mozilla Bug 1400275</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
